New rev of spec incorporating current state of open issues.  Note that this does _not_ yet match the Python code, this is being offered up for comment before making code changes and doing final updates to examples.
Specifically this version changes the signing
algorithm by adding in the MIME type of the
signed entity into the hash-to-sign, per
security recommendations.  The intent is for
this round of changes to be the last set of
signature-breaking changes.

This also adds in multiple signatures since
we're breaking everyone anyway.


git-svn-id: http://salmon-protocol.googlecode.com/svn/trunk@97 0c3e17fe-a321-11de-97f0-97815e7b2756